gumbo-rest-client/pyproject.toml
[tool.poetry]
name = "gumbo-rest-client"
version = "2.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gumbo_rest_client"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pytest = "^7.4.4"
google-auth = "^2.26.2"
pandas = "^1.4"
dataframe-json-packing = { git = "https://github.com/broadinstitute/gumbo_client.git", subdirectory = "dataframe-json-packing" }
# dataframe-json-packing = {path = "../dataframe-json-packing", develop = true}

[tool.poetry.group.dev.dependencies]
gumbo-rest-service = {path = "../gumbo-rest-service", develop = true}
httpx = "^0.26.0"
pyright = "^1.1.347"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
